requirement.txt
certifi==2017.11.5
chardet==3.0.4
cycler==0.10.0
Django==1.10
django-embed-video==1.1.2
idna==2.6
matplotlib==2.0.2
numpy==1.13.3
olefile==0.44
Pillow==4.3.0
pkg-resources==0.0.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
requests==2.18.4
six==1.11.0
urllib3==1.22
